+++ include/value.h
@@ -1,11 +1,13 @@
 #ifndef CLOX_VALUE_H
 #define CLOX_VALUE_H
 
+#include <stddef.h>
+
 typedef double Value;
 
 typedef struct {
-    int capacity;
-    int count;
+    size_t capacity;
+    size_t count;
     Value* values;
 } ValueArray;
